+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add DataGenerators for easier dev
 - Add DataGen for vanilla eggs
 - Recipe Provider for standard and breeding recipes
+- Language provider for the English lang
 
 ### Changed
 - Change build to 1.19.2